+++ config/settings.php
@@ -54,6 +54,7 @@ return [
             'connection' => [
                 'driver'   => 'pdo_mysql',
                 'host'     => '127.0.0.1',
+                'port'     => 3306,
                 'dbname'   => '',
                 'user'     => '',
                 'password' => '',